+++ backend/lib/edgehog_web/resolvers/devices.ex
@@ -1,7 +1,7 @@
 #
 # This file is part of Edgehog.
 #
-# Copyright 2021 SECO Mind Srl
+# Copyright 2021-2023 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,6 +25,7 @@ defmodule EdgehogWeb.Resolvers.Devices do
   alias Edgehog.Devices.SystemModel
   alias Edgehog.Labeling.DeviceAttribute
   alias EdgehogWeb.Schema.VariantTypes
+  alias I18nHelpers.Ecto.Translator
 
   @device_fields_querying_astarte [
     :capabilities,
@@ -41,10 +42,9 @@ defmodule EdgehogWeb.Resolvers.Devices do
     :network_interfaces
   ]
 
-  def find_device(%{id: id}, %{context: context} = resolution) do
+  def find_device(%{id: id}, resolution) do
     device =
       Devices.get_device!(id)
-      |> preload_localized_system_model(context)
       |> maybe_preload_astarte_resources_for_device(resolution)
 
     {:ok, device}
@@ -68,47 +68,31 @@ defmodule EdgehogWeb.Resolvers.Devices do
     {:ok, part_numbers}
   end
 
-  def list_devices(_parent, %{filter: filter}, %{context: context} = resolution) do
+  def list_devices(_parent, %{filter: filter}, resolution) do
     devices =
       Devices.list_devices(filter)
-      |> preload_localized_system_model(context)
       |> maybe_preload_astarte_resources_for_device(resolution)
 
     {:ok, devices}
   end
 
-  def list_devices(_parent, _args, %{context: context} = resolution) do
+  def list_devices(_parent, _args, resolution) do
     devices =
       Devices.list_devices()
-      |> preload_localized_system_model(context)
       |> maybe_preload_astarte_resources_for_device(resolution)
 
     {:ok, devices}
   end
 
-  def update_device(%{device_id: id} = attrs, %{context: context} = resolution) do
+  def update_device(%{device_id: id} = attrs, resolution) do
     device = Devices.get_device!(id)
     attrs = maybe_wrap_typed_values(attrs)
 
     with {:ok, device} <- Devices.update_device(device, attrs) do
-      device =
-        device
-        |> preload_localized_system_model(context)
-        |> maybe_preload_astarte_resources_for_device(resolution)
+      device = maybe_preload_astarte_resources_for_device(device, resolution)
 
       {:ok, %{device: device}}
     end
-  end
-
-  def preload_localized_system_model(target, context) do
-    descriptions_query =
-      context
-      |> Map.fetch!(:preferred_locales)
-      |> Devices.localized_system_model_description_query()
-
-    preload = [descriptions: descriptions_query, hardware_type: [], part_numbers: []]
-
-    Devices.preload_system_model(target, preload: preload)
   end
 
   defp maybe_preload_astarte_resources_for_device(device, resolution) do
@@ -152,24 +136,14 @@ defmodule EdgehogWeb.Resolvers.Devices do
     end
   end
 
-  def find_system_model(%{id: id}, %{context: context}) do
-    with {:ok, system_model} <- Devices.fetch_system_model(id) do
-      system_model = localize_system_model_description(system_model, context)
-      {:ok, system_model}
-    end
+  def find_system_model(%{id: id}, _resolution) do
+    Devices.fetch_system_model(id)
   end
 
-  def list_system_models(_parent, _args, %{context: context}) do
-    system_models =
-      Devices.list_system_models()
-      |> localize_system_model_description(context)
+  def list_system_models(_parent, _args, _resolution) do
+    system_models = Devices.list_system_models()
 
     {:ok, system_models}
-  end
-
-  defp localize_system_model_description(target, context) do
-    %{preferred_locales: preferred_locales} = context
-    Devices.preload_localized_descriptions_for_system_model(target, preferred_locales)
   end
 
   def extract_system_model_part_numbers(
@@ -182,42 +156,32 @@ defmodule EdgehogWeb.Resolvers.Devices do
     {:ok, part_numbers}
   end
 
-  def create_system_model(_parent, %{hardware_type_id: hw_type_id} = attrs, %{
-        context: %{current_tenant: current_tenant} = context
-      }) do
-    default_locale = current_tenant.default_locale
+  def create_system_model(_parent, %{hardware_type_id: hw_type_id} = attrs, resolution) do
+    default_locale = resolution.context.current_tenant.default_locale
 
     with {:ok, hardware_type} <- Devices.fetch_hardware_type(hw_type_id),
          :ok <- check_description_locale(attrs[:description], default_locale),
          attrs = wrap_description(attrs),
          {:ok, system_model} <-
            Devices.create_system_model(hardware_type, attrs) do
-      system_model = localize_system_model_description(system_model, context)
-
       {:ok, %{system_model: system_model}}
     end
   end
 
-  def update_system_model(_parent, %{system_model_id: id} = attrs, %{
-        context: %{current_tenant: current_tenant} = context
-      }) do
-    default_locale = current_tenant.default_locale
+  def update_system_model(_parent, %{system_model_id: id} = attrs, resolution) do
+    default_locale = resolution.context.current_tenant.default_locale
 
     with {:ok, %SystemModel{} = system_model} <- Devices.fetch_system_model(id),
          :ok <- check_description_locale(attrs[:description], default_locale),
          attrs = wrap_description(attrs),
-         system_model = localize_system_model_description(system_model, context),
          {:ok, %SystemModel{} = system_model} <-
            Devices.update_system_model(system_model, attrs) do
-      system_model = localize_system_model_description(system_model, context)
-
       {:ok, %{system_model: system_model}}
     end
   end
 
-  def delete_system_model(%{system_model_id: id}, %{context: context}) do
+  def delete_system_model(%{system_model_id: id}, _resolution) do
     with {:ok, %SystemModel{} = system_model} <- Devices.fetch_system_model(id),
-         system_model = localize_system_model_description(system_model, context),
          {:ok, %SystemModel{} = system_model} <- Devices.delete_system_model(system_model) do
       {:ok, %{system_model: system_model}}
     end
@@ -228,32 +192,32 @@ defmodule EdgehogWeb.Resolvers.Devices do
   defp check_description_locale(%{locale: default_locale}, default_locale), do: :ok
   defp check_description_locale(%{locale: _other}, _default), do: {:error, :not_default_locale}
 
-  # If it's there, wraps description to descriptions, as {create,update}_system_model expects a list
-  defp wrap_description(%{description: description} = attrs) do
-    descriptions = Enum.reject([description], &(is_nil(&1) || String.trim(&1.text) == ""))
-
-    attrs
-    |> Map.delete(:description)
-    |> Map.put(:descriptions, descriptions)
+  # If it's there, wraps description in a map, as {create,update}_system_model expect a map
+  defp wrap_description(%{description: description} = attrs) when is_map(description) do
+    %{locale: locale, text: text} = description
+    %{attrs | description: %{locale => text}}
   end
 
   defp wrap_description(attrs), do: attrs
 
-  def extract_localized_description(%SystemModel{descriptions: descriptions}, _args, %{
-        context: %{preferred_locales: preferred_locales}
-      })
-      when is_list(descriptions) do
-    locale_to_index_map =
-      preferred_locales
-      |> Enum.with_index()
-      |> Enum.into(%{})
+  def extract_localized_description(%SystemModel{} = system_model, _args, resolution) do
+    %{
+      context: %{
+        preferred_locales: preferred_locales,
+        tenant_locale: tenant_locale
+      }
+    } = resolution
 
-    description =
-      descriptions
-      |> Enum.sort_by(&Map.get(locale_to_index_map, &1.locale))
-      |> List.first()
+    # TODO: move this in a middleware
+    %SystemModel{translated_description: translated_description} =
+      Translator.translate(system_model, preferred_locales, fallback_locale: tenant_locale)
 
-    {:ok, description}
+    # TODO: fix the library to return nil on empty translations
+    if translated_description == "" do
+      {:ok, nil}
+    else
+      {:ok, translated_description}
+    end
   end
 
   def extract_device_tags(%Device{tags: tags}, _args, _context) do