+++ backend/test/edgehog_web/schema/query/base_image_collections_test.exs
@@ -1,7 +1,7 @@
 #
 # This file is part of Edgehog.
 #
-# Copyright 2022 SECO Mind Srl
+# Copyright 2022-2023 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -40,10 +40,7 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
         name
         handle
         systemModel {
-          description {
-            locale
-            text
-          }
+          description
         }
       }
     }
@@ -98,12 +95,12 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
     } do
       default_locale = tenant.default_locale
 
-      descriptions = [
-        %{locale: default_locale, text: "A base image collection"},
-        %{locale: "it-IT", text: "Un modello di sistema"}
-      ]
+      description = %{
+        default_locale => "A base image collection",
+        "it-IT" => "Un modello di sistema"
+      }
 
-      system_model = system_model_fixture(hardware_type, descriptions: descriptions)
+      system_model = system_model_fixture(hardware_type, description: description)
 
       _base_image_collection = base_image_collection_fixture(system_model)
 
@@ -114,10 +111,7 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
                  "baseImageCollections" => [
                    %{
                      "systemModel" => %{
-                       "description" => %{
-                         "locale" => ^default_locale,
-                         "text" => "A base image collection"
-                       }
+                       "description" => "A base image collection"
                      }
                    }
                  ]
@@ -133,12 +127,12 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
     } do
       default_locale = tenant.default_locale
 
-      descriptions = [
-        %{locale: default_locale, text: "A base image collection"},
-        %{locale: "it-IT", text: "Un modello di sistema"}
-      ]
+      description = %{
+        default_locale => "A base image collection",
+        "it-IT" => "Un modello di sistema"
+      }
 
-      system_model = system_model_fixture(hardware_type, descriptions: descriptions)
+      system_model = system_model_fixture(hardware_type, description: description)
 
       _base_image_collection = base_image_collection_fixture(system_model)
 
@@ -152,10 +146,7 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
                  "baseImageCollections" => [
                    %{
                      "systemModel" => %{
-                       "description" => %{
-                         "locale" => "it-IT",
-                         "text" => "Un modello di sistema"
-                       }
+                       "description" => "Un modello di sistema"
                      }
                    }
                  ]
@@ -171,12 +162,12 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
     } do
       default_locale = tenant.default_locale
 
-      descriptions = [
-        %{locale: default_locale, text: "A base image collection"},
-        %{locale: "it-IT", text: "Un modello di sistema"}
-      ]
+      description = %{
+        default_locale => "A base image collection",
+        "it-IT" => "Un modello di sistema"
+      }
 
-      system_model = system_model_fixture(hardware_type, descriptions: descriptions)
+      system_model = system_model_fixture(hardware_type, description: description)
 
       _base_image_collection = base_image_collection_fixture(system_model)
 
@@ -190,10 +181,7 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
                  "baseImageCollections" => [
                    %{
                      "systemModel" => %{
-                       "description" => %{
-                         "locale" => ^default_locale,
-                         "text" => "A base image collection"
-                       }
+                       "description" => "A base image collection"
                      }
                    }
                  ]
@@ -207,11 +195,11 @@ defmodule EdgehogWeb.Schema.Query.BaseImageCollectionsTest do
            api_path: api_path,
            hardware_type: hardware_type
          } do
-      descriptions = [
-        %{locale: "it-IT", text: "Un modello di sistema"}
-      ]
+      description = %{
+        "it-IT" => "Un modello di sistema"
+      }
 
-      system_model = system_model_fixture(hardware_type, descriptions: descriptions)
+      system_model = system_model_fixture(hardware_type, description: description)
 
       _base_image_collection = base_image_collection_fixture(system_model)